Fix module installation
When redCORE is installed from a extension the mod_redcore_language_switcher is not installed because is not found by the redCORE installer.

This fix uses a feature already provided in the installer https://github.com/redCOMPONENT-COM/redCORE/blob/develop/install.php#L287 setting the right client. This fix solves all current issues when installing redCORE from an extension.

Still this fix is going to be improved with a future pull that will provide a common script for all extensions