Web service fix
This change allows to use custom named functions inside the model for
getting web service results without any problems. Main problem with
current solution is in getting pagination for results. Current hal use
model `getPagination` function, which can't apply on any custom function
in a model for getting results.
With this change, we are not using any pagination for any custom
function, so this is up to a developer to make it right for his own
custom functions.